fix #826: import GPX from email/browser without GPXImportActivity
- import from email/browser is handled via cgeocaches activity
  (offline list)
